change: int to size_t and %i to %u

+++ singly_linked_lists/0-print_list.c
@@ -11,14 +11,14 @@
 size_t print_list(const list_t *h)
 
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h)
 	{
 		if (h->str == NULL)
 			printf("[0] (nil)\n");
 		else
-			printf("[%i] [%s]\n", h->len, h->str);
+			printf("[%u] [%s]\n", h->len, h->str);
 
 		h = h->next;
 		count++;
